Add sleep and array check to avoid stale file handles due to deletion

+++ JobSubmission/5_batch_CreateIncrementalModels.sh
@@ -204,9 +204,14 @@ finishing_statement 1; }
 output_directory="${MODEL_DIR:?}\
 /BinSize_${bin_size}_SampleSize_${sample_size}_${number_of_models_to_generate}"
 
-rm -rf "${output_directory}"
-mkdir -p "${output_directory}/Likelihood_Values"
-mkdir -p "${output_directory}/STATEBYLINE"
+sleep 5
+# Avoid stale file handles
+if [[ "${SLURM_ARRAY_TASK_ID}" -eq 1 ]]; then
+    rm -rf "${output_directory}"
+    mkdir -p "${output_directory}/Likelihood_Values"
+    mkdir -p "${output_directory}/STATEBYLINE"
+fi
+
 cd "${output_directory}/Likelihood_Values" || finishing_statement 1
